fix: Add `patch` to requirements documentation
I noticed that `patch` was not installed on my system, leading to Electron being
imported in the backend, which then caused an error: "Electron failed to install
correctly, please delete node_modules/electron and try installing again". The
build steps are currently written such that they succeed even if the patch
command fails. This works for now since it makes them idempotent, since patch
fails if it cannot apply changes (e.g. the changes are already applied).
However, it also hides the case where patch isn't installed. I'd like to ensure
patch is installed and provide an error otherwise, but for now we can just
update the documentation to let users know it's required.