fix: Revert github runners for E2E tests
Use the github provided runners for now because the E2E tests are
expensive to run and it's unclear how much value the larger runner
provides. Tor seems to cause a lot of variability in test run time. We
are still using the larger Linux runner for the Android build.