fix: Clean up desktop UI console errors/warnings
On desktop, the browser was emitting several errors/warnings that
filled up the console. Some of these were related to React and then
some of them related to HTML preload. The preload warning says, "A
preload for {url} is found, but is not used because the request
credentials mode does not match. Consider taking a look at crossorigin
attribute." and then a little later it says, "The resource {url} was
preloaded using link preload but not used within a few seconds from
the window's load event. Please make sure it has an appropriate `as`
value and it is preloaded intentionally." Removing the crossorigin
attribute on the font links removed the preload warning and it looks
like everything is preloading correctly to me in dev/production (fonts
load at the very beginning and no warnings). Maybe if CORS is not used
then it also doesn't want crossorigin attribute.