feat: Add user profile feature for desktop (#1923)
On the frontend, the user profile sidebar panel simply shows the current user's
name and profile photo. Clicking on the profile photo brings up a context menu
that shows the profile and profile actions (only 'edit profile' for now).
Clicking the 'edit profile' action changes the context menu view such that you
can edit the profile photo. This commit also adds user profile photos to channel
messages. A jdenticon is showed in place of user profile photo if it doesn't
exist. Message profile photos also show on mobile.

In state-manager, a new saveUserProfile saga is added which signs the profile
data and sends it to the backend.

On the backend, there is a new shared OrbitDB user profiles key/value store. The
key/value store maps user public keys with profiles. It uses a custom index so
that we can filter invalid entries thoroughly. Profile photos are stored in OrbitDB
entries for now.