refactor: Move community creation and data to backend (#2305)
This commit migrates the Community model to the backend so that the backend is now the source of truth for Community data. In order to accomplish this, for existing communities (existing Quiet install), we move (migrate) the data from the frontend to the backend and then make sure all operations that previously affected the frontend Community state are done on the backend. For new communities (new Quiet install), some of the logic for creating/joining has been moved to the backend and the event-flow is simplified. Whenever the Community model is updated on the backend, the backend now emits a COMMUNITY_UPDATED event so that the frontend can synchronize its Community state.

Solution for: #2329