Podfile changes to get RN 74 working on iOS and
Fix React Native 0.74 iOS broken bundleUrl in AppDelegate.m

https://reactnative.dev/blog/2024/04/22/release-0.74#other-breaking-changes
On iOS prior to RN 0.74 Before, bundleURL was set when React Native was started in an instance variable and it was not possible to update it.
Now, bundleUrl is a function which is re-evaluated when needed, enabling the use of a different URL across refreshes.

https://stackoverflow.com/questions/78582786/rctappdelegatebundleurl-not-implemented-reason-subclasses-must-implement-a